+++ src/mangle.h
@@ -32,21 +32,21 @@
  * Processes the packet and returns verdict.
  * This is the primary function that traverses the packet.
  */
-int process_packet(const uint8_t *packet, uint32_t packet_len);
+int process_packet(const uint8_t *packet, size_t packet_len);
 
 
 /**
  * Processe the TCP packet.
  * Returns verdict.
  */
-int process_tcp_packet(const struct section_config_t *section, const uint8_t *raw_payload, uint32_t raw_payload_len);
+int process_tcp_packet(const struct section_config_t *section, const uint8_t *raw_payload, size_t raw_payload_len);
 
 
 /**
  * Processes the UDP packet.
  * Returns verdict.
  */
-int process_udp_packet(const struct section_config_t *section, const uint8_t *pkt, uint32_t pktlen);
+int process_udp_packet(const struct section_config_t *section, const uint8_t *pkt, size_t pktlen);
 
 
 
@@ -63,8 +63,8 @@ int post_fake_sni(struct fake_type f_type,
  * dvs used internally and should be zero.
  */
 int send_tcp_frags(const struct section_config_t *section,
-	const uint8_t *packet, uint32_t pktlen, 
-	const uint32_t *poses, uint32_t poses_len, uint32_t dvs);
+	const uint8_t *packet, size_t pktlen, 
+	const size_t *poses, size_t poses_len, size_t dvs);
 
 /**
  * Splits packet by poses and posts.
@@ -72,6 +72,6 @@ int send_tcp_frags(const struct section_config_t *section,
  * dvs used internally and should be zero.
  */
 int send_ip4_frags(const struct section_config_t *section,
-	const uint8_t *packet, uint32_t pktlen, 
-	const uint32_t *poses, uint32_t poses_len, uint32_t dvs);
+	const uint8_t *packet, size_t pktlen, 
+	const size_t *poses, size_t poses_len, size_t dvs);
 #endif /* YU_MANGLE_H */
